Data

+++ db/migrate/20230530084623_create_cars.rb
@@ -6,7 +6,7 @@ class CreateCars < ActiveRecord::Migration[7.0]
       t.string :description
       t.integer :rating
       t.decimal :price, precision: 10, scale: 2
-      t.integer :rent_per_day
+      t.decimal :rent_per_day
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
